requirements.txt
-e git+http://github.com/django/django.git@1.7b3#egg=Django
django-admin-bootstrapped==1.6.3
xlwt==0.7.5
gunicorn==18.0
#psycopg2==2.5.1
#dj-database-url==0.2.2
#dj-static==0.0.5
#static==0.4
wsgiref==0.1.2
django_easy_select2==1.2.2
django_extensions
xlrd
